% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cetesb_aqs}
\alias{cetesb_aqs}
\title{CETESB AQS stations.}
\format{
A data frame with two variables: \code{name} and \code{code}
of AQS.
}
\usage{
cetesb_aqs
}
\description{
List of CETESB air quality stations (AQS) with their codes in
CETESB QUALAR system. Use this to check the aqs_code parameter.
AQS names are without diacritics.
}
\keyword{datasets}
