% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cetesb_param}
\alias{cetesb_param}
\title{CETESB Parameters}
\format{
A data frame with 20 observations and 3 variables:
\describe{
\item{name}{CETESB QUALAR paramater abbreviation and name.}
\item{units}{Parameter units.}
\item{code}{Parameter CETESB QUALAR code.}
}
}
\usage{
cetesb_param
}
\description{
List of CETESB QUALAR available parameters and units. Use this to check the
\code{parameters} argument.
Parameter names are without diacritics.
}
\examples{
cetesb_param
}
\keyword{datasets}
