% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CetesbRetrieve.R
\name{CetesbRetrieve}
\alias{CetesbRetrieve}
\title{Download a single parameter from CETESB QUALAR system}
\usage{
CetesbRetrieve(username, password, pol_code, aqs_code, start_date, end_date)
}
\arguments{
\item{username}{User name of CETESB QUALAR}

\item{password}{User name's password of CETESB QUALAR}

\item{pol_code}{Code of parameter}

\item{aqs_code}{Code of AQS}

\item{start_date}{Date to start downloading in \%dd/\%mm/\%yyyy}

\item{end_date}{Date to end downloading in \%dd/\%mm/\%yyyy}
}
\value{
data.frame with the selected parameter information
}
\description{
This function download one parameter from one air quality stations (AQS)
of CETESB AQS network. It will pad out the date with missing data with NA.
}
\examples{
\dontrun{
# Downloading Ozone information from Pinheiros AQS
# from January first to 7th of 2020
my_user_name <- "John Doe"
my_pass_word <- "drowssap"
o3_code <- 63  # Check with cetesb_param
pin_code <- 99 # Check with cetesb_aqs
start_date <- "01/01/2020"
end_date <- "07/01/2020"

pin_o3 <- CetesbRetrieve(my_user_name, my_pass_word, o3_code, pin_code, start_date, end_date)

}
}
