% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CetesbRetrieveMetPol.R
\name{CetesbRetrieveMetPol}
\alias{CetesbRetrieveMetPol}
\title{Download meteorological and pollutant data from CETESB QUALAR}
\usage{
CetesbRetrieveMetPol(username, password, aqs_code, start_date, end_date)
}
\arguments{
\item{username}{User name of CETESB QUALAR}

\item{password}{User name's password of CETESB QUALAR}

\item{aqs_code}{Code of AQS}

\item{start_date}{Date to start downloading in dd/mm/yyyy}

\item{end_date}{Date to end downloading in dd/mm/yyyy}
}
\value{
data.frame with ith Temperature (C), Relative Humidity (\%), Wind Speed (m/s) and Direction (degrees),
Pressure information (hPa), O3, NO, NO2, NOx, PM2.5, PM10 and CO information.
}
\description{
This function download the main meteorological parameters for model evaluation,
together with criteria pollutants for one air quality station (AQS).
It will pad out the date with missing data with NA.
}
\examples{
\dontrun{
# Downloading main meteorological parameters and criteria pollutants from Pinheiros AQS
# from January first to 7th of 2020
my_user_name <- "John Doe"
my_pass_word <- "drowssap"
pin_code <- 99 # Check with cetesb_aqs
start_date <- "01/01/2020"
end_date <- "07/01/2020"

pin_pol <- CetesbRetrieveMetPol(my_user_name, my_pass_word, pin_code, start_date, end_date)

}
}
