% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckCetesbCode.R
\name{CheckCetesbCode}
\alias{CheckCetesbCode}
\title{Internal function - Check if CETESB AQS code or name is valid}
\usage{
CheckCetesbCode(aqs, aqs_code)
}
\arguments{
\item{aqs}{data frame with cetesb aqs code and names}

\item{aqs_code}{user's input code or name}
}
\description{
Internal function - Check if CETESB AQS code or name is valid
}
\keyword{internal}
