% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monitor_ar_aqs}
\alias{monitor_ar_aqs}
\title{Monitor Ar AQS stations.}
\format{
A data frame with 8 observation and 6 variables:
\describe{
\item{name}{MonitorAr Program AQS name.}
\item{code}{MonitorAr Program AQS abbreviation.}
\item{lon}{MonitorAr Program AQS longitude.}
\item{lat}{MonitorAr Program AQS latitude.}
\item{X_UTM_Sirgas2000}{MonitorAr Program AQS longitude in SIRGAS 2000
(EPSG:31983).}
\item{Y_UTM_Sirgas2000}{MonitorAr Program AQS latitude in SIRGAS 2000
(EPSG:31983).}
}
}
\usage{
monitor_ar_aqs
}
\description{
List of Monitor Ar Rio air quality stations (AQS) with their codes
and locations. Use this to check the parameters argument in
MonitorArRetrieveParam() function.
}
\examples{
monitor_ar_aqs
}
\keyword{datasets}
