% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monitor_ar_param}
\alias{monitor_ar_param}
\title{Monitor Ar Parameters}
\format{
A data frame with 3 variables: \code{code}, \code{name}, and \code{units}
of parameter.
}
\usage{
monitor_ar_param
}
\description{
List of Monitor Ar Rio available parameters. Use this to check the
param parameter of MonitorArRetrieve() function .
}
\keyword{datasets}
