% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ar_retrieve_met.R
\name{monitor_ar_retrieve_met}
\alias{monitor_ar_retrieve_met}
\title{Download meteorological parameters from Monitor Ar program}
\usage{
monitor_ar_retrieve_met(
  start_date,
  end_date,
  aqs_code,
  to_local = TRUE,
  to_csv = FALSE,
  csv_path = ""
)
}
\arguments{
\item{start_date}{Date to start downloading in dd/mm/yyyy.}

\item{end_date}{Date to end downloading in dd/mm/yyyy.}

\item{aqs_code}{Code of AQS. See \code{monitor_ar_aqs}.}

\item{to_local}{Date information in local time. TRUE by default.}

\item{to_csv}{Create a csv file. FALSE by default.}

\item{csv_path}{Path to save the csv file.}
}
\value{
data.frame with Temperature (c), Relative Humidity (\%),
Wind speed (m/s) and direction (degrees) and Pressure information.
}
\description{
This function download the main meteorological parameters from one
air quality station (AQS) of Monitor Ar network. It will
pad out the date with missing data with NA.
}
\examples{
\dontrun{
# Downloading meteorological data from CENTRO AQS
# from January first to 7th of 2020
start_date <- "01/01/2020"
end_date <- "07/01/2020"
ca_met <- monitor_ar_retrieve_met(start_date, end_date, "CA")
}
}
