% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ar_retrieve_pol.R
\name{monitor_ar_retrieve_pol}
\alias{monitor_ar_retrieve_pol}
\title{Download criteria pollutants from Monitor Ar program}
\usage{
monitor_ar_retrieve_pol(
  start_date,
  end_date,
  aqs_code,
  to_local = TRUE,
  to_csv = FALSE,
  csv_path = ""
)
}
\arguments{
\item{start_date}{Date to start downloading in dd/mm/yyyy.}

\item{end_date}{Date to end downloading in dd/mm/yyyy.}

\item{aqs_code}{Code of AQS.}

\item{to_local}{Date information in local time. TRUE by default.}

\item{to_csv}{Creates a csv file. FALSE by default.}

\item{csv_path}{Path to save the csv file.}
}
\value{
data.frame with O3, NO, NO2, NOx, PM2.5, PM10 and CO
information.
}
\description{
This function download the criteria pollutants from
one air quality station (AQS) of Monitor Ar Program.
It will pad out the date with missing data with NA.
}
\examples{
\dontrun{
# Downloading criteria pollutants from CENTRO AQS
# from January first to 7th of 2020
start_date <- "01/01/2020"
end_date <- "07/01/2020"
ca_pol <- monitor_ar_retrieve_pol(start_date, end_date, "CA")
}
}
