% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorArRetrieveMet.R
\name{MonitorArRetrieveMet}
\alias{MonitorArRetrieveMet}
\title{Download meteorological parameters from Monitor Ar program}
\usage{
MonitorArRetrieveMet(
  start_date,
  end_date,
  aqs_code,
  verbose = TRUE,
  to_local = TRUE,
  to_csv = FALSE
)
}
\arguments{
\item{start_date}{Date to start downloading in dd/mm/yyyy.}

\item{end_date}{Date to end downloading in dd/mm/yyyy.}

\item{aqs_code}{Code of AQS. See \code{monitor_ar_aqs}.}

\item{verbose}{Print query summary.}

\item{to_local}{Date information in local time. TRUE by default.}

\item{to_csv}{Create a csv file. FALSE by default.}
}
\value{
data.frame with Temperatur (c), Relative Humidity (\%), Wind speed (m/s) and direction (degrees)
and Pressure information
}
\description{
This function download the main meterorological paramenters from one
air quality station (AQS) of Monitor Ar network. It will
pad out the date with missing data with NA.
}
\examples{
\dontrun{
# Downloading meteorological data from CENTRO AQS
# from January first to 7th of 2020
start_date <- "01/01/2020"
end_date <- "07/01/2020"
ca_met <- MonitorArRetrieveMet(start_date, end_date, "CA")
}
}
