% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorArRetrieveMetPol.R
\name{MonitorArRetrieveMetPol}
\alias{MonitorArRetrieveMetPol}
\title{Download meteorological and pollutant data from Monitor Ar Program}
\usage{
MonitorArRetrieveMetPol(
  start_date,
  end_date,
  aqs_code,
  verbose = TRUE,
  to_local = TRUE,
  to_csv = FALSE
)
}
\arguments{
\item{start_date}{Date to start downloading dd/mm/yyyy}

\item{end_date}{Date to end downloading dd/mm/yyyy}

\item{aqs_code}{Code of AQS}

\item{verbose}{Print query summary}

\item{to_local}{Date information in local time. TRUE by default.}

\item{to_csv}{Creates a csv file. FALSE by default.}
}
\value{
data.frame with Temperature (C), Relative Humidity (\%), Wind Speed (m/s) and
Direction (degrees), Pressure information (hPa), O3, NO, NO2, NOx, PM2.5, PM10 and CO
information
}
\description{
This function download the main meteorological parameters for model evaluation,
together with criteria pollutatns for in air quality station (AQS) of Monitor Ar
program. It will pad out the date with missing data with NA
}
\examples{
\dontrun{
# Downloading main meteorological parameters and criteria pollutant from CENTRO AQS
# from January first to 7th of 2020
start_date <- "01/01/2020"
end_date <- "07/01/2020"
ca_all <- MonitorArRetrieceMetPol(start_date, end_date, "CA")
}
}
