% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorArRetrieveParam.R
\name{MonitorArRetrieveParam}
\alias{MonitorArRetrieveParam}
\title{Download air quality and meteorology information from MonitorAr-Rio}
\usage{
MonitorArRetrieveParam(
  start_date,
  end_date,
  aqs_code,
  param,
  to_local = TRUE,
  verbose = TRUE,
  to_csv = FALSE
)
}
\arguments{
\item{start_date}{Date to start downloading in dd/mm/yyyy}

\item{end_date}{Date to end downloading in dd/mm/yyyy}

\item{aqs_code}{AQS code}

\item{param}{Paremeter to download. It can be a vector with many parameters.}

\item{to_local}{Date information in local time. TRUE by default.}

\item{verbose}{Print query summary.}

\item{to_csv}{Creates a csv file. FALSE by default}
}
\value{
data.frame with the selected parameter information
}
\description{
This function download air quality and meteorology measurements from MonitorAr-Rio
program from Rio de Janeiro city.
}
\examples{
\dontrun{
# Downloading Ozone information from Centro AQS
# from February of 2019
date_start <- "01/02/2019"
date_end <- "01/03/2019"
aqs_code <- "CA"
param <- "O3"
ca_o3 <- MonitorArRetrieve(date_start, date_end, aqs_code, param)

}
}
