% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteCSV.R
\name{WriteCSV}
\alias{WriteCSV}
\title{Internal function - Export to csv data download}
\usage{
WriteCSV(aqs_data_df, aqs_name, start_date, end_date, vars)
}
\arguments{
\item{aqs_name}{AQS name}

\item{start_date}{Download start date}

\item{end_date}{Download end date}

\item{vars}{identifier in file name}

\item{aqs_data_frame}{Downloaded dataframe}
}
\description{
Internal function - Export to csv data download
}
\keyword{internal}
