% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cetesb_aqs}
\alias{cetesb_aqs}
\title{CETESB AQS station latitude and longitude}
\format{
A data frame with 4 variables: \code{name} , \code{code}, \code{lat}, and
\code{latitude} of AQS.
}
\usage{
cetesb_aqs
}
\description{
List of CETESB QUALAR AQS latitudes and longitudes. Use this to check the
AQS location, if you need to make some maps. AQS names contains diacritics.
}
\keyword{datasets}
