% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cetesb_aqs}
\alias{cetesb_aqs}
\title{CETESB AQS station latitude and longitude}
\format{
A data frame with 68 observations and 4 variables:
\describe{
\item{name}{CETESB AQS name.}
\item{code}{CETESB AQS code in QUALAR System.}
\item{lat}{CETESB AQS latitude.}
\item{lon}{CETESB AQS lontgitude.}
}
}
\usage{
cetesb_aqs
}
\description{
List of CETESB QUALAR air qualitty stations (AQS) latitudes and longitudes.
Use this to check the AQS \code{aqs_code} argument in CetesbRetrieveParam()
function. AQS names are without diacritics.
}
\examples{
cetesb_aqs
}
\keyword{datasets}
