% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cetesb_param}
\alias{cetesb_param}
\title{CETESB Parameters}
\format{
A data frame with 3 variables: \code{name}, \code{units}, and \code{code}
of parameter.
}
\usage{
cetesb_param
}
\description{
List of CETESB QUALAR available parameters and units. Use this to check the
pol_code parameter.
Parameter names are without diacritics.
}
\keyword{datasets}
