% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb_retrieve_pol.R
\name{cetesb_retrieve_pol}
\alias{cetesb_retrieve_pol}
\title{Download criteria pollutants from CETESB QUALAR}
\usage{
cetesb_retrieve_pol(
  username,
  password,
  aqs_code,
  start_date,
  end_date,
  verbose = TRUE,
  to_csv = FALSE,
  csv_path = ""
)
}
\arguments{
\item{username}{User name of CETESB QUALAR}

\item{password}{User name's password of CETESB QUALAR}

\item{aqs_code}{Code of AQS}

\item{start_date}{Date to start downloading in dd/mm/yyyy}

\item{end_date}{Date to end downloading in dd/mm/yyyy}

\item{verbose}{Print query summary}

\item{to_csv}{Creates a csv file. FALSE by default}

\item{csv_path}{Path to save the csv file}
}
\value{
data.frame with O3, NO, NO2, PM2.5, PM10 and CO information.
Units are ug/m3 except for CO which is in ppm, and NOx which is in ppb.
}
\description{
This function download the criteria pollutants from one air quality station
(AQS) of CETESB AQS network.
It will pad out the date with missing data with NA.
This function required to have
\href{https://seguranca.cetesb.sp.gov.br/Home/CadastrarUsuario}{an account}
in \href{https://qualar.cetesb.sp.gov.br/qualar/home.do}{CETESB QUALAR}.
}
\examples{
\dontrun{
# Downloading criteria pollutants from Pinheiros AQS
# from January first to 7th of 2020
my_user_name <- "John Doe"
my_pass_word <- "drowssap"
pin_code <- 99 # Check with cetesb_aqs
start_date <- "01/01/2020"
end_date <- "07/01/2020"

pin_pol <- cetesb_retrieve_pol(my_user_name, my_pass_word, pin_code,
                               start_date, end_date)

}
}
