% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CetesbRetrieveParam.R
\name{CetesbRetrieveParam}
\alias{CetesbRetrieveParam}
\title{Download list of observation from CETESB QUALAR}
\usage{
CetesbRetrieveParam(
  username,
  password,
  parameters,
  aqs_code,
  start_date,
  end_date,
  verbose = TRUE,
  to_csv = FALSE
)
}
\arguments{
\item{username}{User name of CETESB QUALAR}

\item{password}{User name's password of CETESB QUALAR}

\item{parameters}{a character vector with the parameters abbreviations
to download}

\item{aqs_code}{Code of AQS}

\item{start_date}{Date to start downloading in dd/mm/yyyy}

\item{end_date}{Date to end downloading in dd/mm/yyyy}

\item{verbose}{Print query summary}

\item{to_csv}{Creates a csv file. FALSE by default}
}
\value{
data.frame with parameters described in \code{params} vector
}
\description{
This function downloads the parameters in a vector, this parameters
can be both pollutants or meteorological observations for one air
quality station (AQS). It will pad out the date with missing data
}
\examples{
\dontrun{
# Download ozone, nitrogen dioxide, and wind speed
# from Pinheiros AQS, from January first to 7th of 2020

my_user_name <- "John Doe"
my_pass_word <- "drowssap"
pin_code <- 99 # Check with cetesb_aqs
start_date <- "01/01/2020"
end_date <- "07/01/2020"
params <- c("o3","NOX", "VV")

pin_param <- CetesbRetrieveParam(my_user_name, my_pass_word,
                                 params, pin_code,
                                 start_date, end_date)
}
}
