% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monitor_ar_aqs}
\alias{monitor_ar_aqs}
\title{Monitor Ar AQS stations.}
\format{
A data frame with 6 variables: \code{name}, \code{code}, \code{lon}, \code{lat},
\code{X_UTM_Sirgas2000}, and  \code{Y_UTM_Sirgas2000}
of AQS.
}
\usage{
monitor_ar_aqs
}
\description{
List of Monitor Ar Rio air quality stations (AQS) with their codes
and locations. Use this to check the aqs_code parameter in
MonitorArRetrieve() function.
}
\keyword{datasets}
