% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monitor_ar_param}
\alias{monitor_ar_param}
\title{Monitor Ar Parameters}
\format{
A data frame with 18 observations and 3 variables:
\describe{
\item{code}{MonitorAr parameter abbreviation or code.}
\item{name}{MonitorAr parameter name}
\item{units}{Parameter units.}
}
}
\usage{
monitor_ar_param
}
\description{
List of Monitor Ar Rio available parameters. Use this to check the
parameters argument in MonitorArRetrieveParam() function.
Parameter names are without diacritics.
}
\examples{
monitor_ar_param
}
\keyword{datasets}
