% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualR.R
\docType{package}
\name{qualR}
\alias{qualR-package}
\alias{qualR}
\title{qualR: An R package to download Sao Paulo and Rio de Janeiro
air pollution data}
\description{
A package to download information from CETESB QUALAR \url{https://cetesb.sp.gov.br/ar/qualar/} and MonitorAr \url{http://jeap.rio.rj.gov.br/je-metinfosmac/institucional/index.html} systems. It contains function to download different parameters, a set of criteria pollutants and the most frequent meteorological parameters used in air quality data analysis and air quality model evaluation.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository:
\url{https://github.com/ropensci/qualR}}
\item{Report bugs:
\url{https://github.com/ropensci/qualR/issues}}
\item{qualR site:
\url{https://docs.ropensci.org/qualR/}}
}
}
\author{
Mario Gavidia-Calderon and Daniel Schuch and Maria de Fatima Andrade
}
